refactor: utils (#20)
