+++ src/fetchData/getProfileData.ts
@@ -2,7 +2,7 @@
  * @see https://github.com/vercel/next.js/tree/canary/examples/with-firebase
  */
 
-import admin from "../firebase/nodeApp";
+import admin from "@/utils/firebase/nodeApp";
 
 export const getProfileData = async (username: string) => {
   const db = admin.firestore();
